make `sandbox.svelte`
